change Position values to long

+++ java/src/main/java/ch/neukom/advent2024/day6/Part2.java
@@ -1,8 +1,8 @@
 package ch.neukom.advent2024.day6;
 
-import ch.neukom.advent2024.util.inputreaders.InputArrayReader;
 import ch.neukom.advent2024.util.data.Direction;
 import ch.neukom.advent2024.util.data.Position;
+import ch.neukom.advent2024.util.inputreaders.InputArrayReader;
 
 import java.io.IOException;
 import java.util.HashSet;
@@ -63,7 +63,7 @@ public class Part2 {
         while (guardPosition.isInside(width, height)) {
             Position nextPosition = guardPosition.move(guardDirection);
             if (nextPosition.isInside(width, height)) {
-                Boolean hasObstacle = facilityMap[nextPosition.y()][nextPosition.x()];
+                Boolean hasObstacle = facilityMap[(int) nextPosition.y()][(int) nextPosition.x()];
                 if (hasObstacle) {
                     guardDirection = guardDirection.turnRight();
                 } else {
