+++ java/src/main/java/ch/neukom/advent2024/day6/Util.java
@@ -1,8 +1,8 @@
 package ch.neukom.advent2024.day6;
 
-import ch.neukom.advent2024.util.inputreaders.InputArrayReader;
 import ch.neukom.advent2024.util.data.Direction;
 import ch.neukom.advent2024.util.data.Position;
+import ch.neukom.advent2024.util.inputreaders.InputArrayReader;
 
 import java.util.List;
 
@@ -22,12 +22,12 @@ public class Util {
         while (guardPosition.isInside(width, height)) {
             Position nextPosition = guardPosition.move(guardDirection);
             if (nextPosition.isInside(width, height)) {
-                Boolean hasObstacle = facilityMap[nextPosition.y()][nextPosition.x()];
+                Boolean hasObstacle = facilityMap[(int) nextPosition.y()][(int) nextPosition.x()];
                 if (hasObstacle) {
                     guardDirection = guardDirection.turnRight();
                 } else {
                     guardPosition = nextPosition;
-                    patrolMap[guardPosition.y()][guardPosition.x()] = true;
+                    patrolMap[(int) guardPosition.y()][(int) guardPosition.x()] = true;
                 }
             } else {
                 guardPosition = nextPosition;
